requirements.txt
authlib==1.6.5
b2sdk>=1.22.0
fastapi==0.119.0
numpy==2.2.6
opencv-python==4.12.0.88
passlib==1.7.4
bcrypt>=4.0.0,<4.1.0
pdfplumber==0.11.7
pydantic==2.12.0
pymongo==4.15.3
python-dotenv==1.1.1
python_jose==3.5.0
reportlab==4.2.5
SQLAlchemy==2.0.44
starlette==0.48.0
uvicorn